feat: Amend UpdateRequiredErrorScreen and VM - inject the ApplicationContext instead of passing it into the fun - fix UpdateRequiredErrorScreenTest where testing the intent
Resolves: DCMAW-7426
