+++ app/src/main/java/uk/gov/onelogin/ui/error/update/UpdateRequiredErrorScreen.kt
@@ -3,7 +3,6 @@ package uk.gov.onelogin.ui.error.update
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
 import uk.gov.android.onelogin.R
 import uk.gov.android.ui.components.R as UiR
@@ -16,7 +15,6 @@ import uk.gov.android.ui.theme.smallPadding
 fun UpdateRequiredScreen(
     viewModel: UpdateRequiredErrorViewModel = hiltViewModel()
 ) {
-    val context = LocalContext.current
     LandingPage(
         landingPageParameters = LandingPageParameters(
             topIcon = UiR.drawable.ic_error,
@@ -34,7 +32,7 @@ fun UpdateRequiredScreen(
             ),
             contentInternalPadding = PaddingValues(bottom = smallPadding),
             primaryButtonText = R.string.app_updateAppButton,
-            onPrimary = { viewModel.updateApp(context) }
+            onPrimary = { viewModel.updateApp() }
         )
     )
 }
